fix: remove console log and change quantity parsing to float in SupplyImportFormDialog

+++ src/modules/supplies/components/supply-import-form-dialog.tsx
@@ -497,7 +497,7 @@ export function SupplyImportFormDialog({
                       control={form.control}
                       type="number"
                       onChange={(e) => {
-                        const quantity = parseInt(e.target.value) || 0;
+                        const quantity = parseFloat(e.target.value) || 0;
                         handleQuantityChange(index, quantity);
                       }}
                     />
